requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9
aiosmtplib==3.0.1
pydantic==2.6.1
pyyaml==6.0.1
slowapi==0.1.8
aioredis==2.0.1
redis==5.0.1
email-validator==2.1.0.post1
